requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
coverage==5.3.1
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
PyJWT==2.0.0
pylint==2.6.0
pymongo==3.6.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.2
wrapt==1.12.1
